Calendar event changes: UID add
Added a UID field, which is important for icalendar. Format is a bit hacky:

<timestamp>-<event #>@<hostname>

Event # is just a running count; I don't bother adding this till I write out the file. This is not actually in spec, since UID is supposed to stay the same for any given calendar event. I'll start thinking about fixing this if I ever start tracking events in a database, which would have some advantages but not enough to be a priority for 1.0.

Also snuck in a hardcoded skip for "Rent The Beacon" events, which really are not in scope for upcoming movies.