Added individual theater ICS files
Also refactored make_film_calendar to dynamically load classes, because the repeated code for each theater was getting silly. importlib is easier than I thought it'd be.

In the process I renamed the Beacon's class because ... I'm a purist, the name of the theater is The Beacon, and I felt wrong abbreviating it to just "Beacon."