Adding RSS support -- first pass
This is a roughed out version without perfect formatting; make_film_calendar.py is also just serving as a test harness right now.

To do:

- test in an RSS reader
- fix the date representation
- figure out if there are other formatting changes needed

Of note:

```
            event_time = vDatetime.from_ical(event["dtstart"].to_ical())
```

That line is about as silly as it looks. The icalendar library docs claim that event.decoded("dtstart") should work but there's a bug in python 3.10, so I'm just taking the long way round.