Added config file
Which was chunky and I'm now messing around with the internals of CLick more than I expected. But it's functional even if I'm not sure I'm being perfectly elegant here.

Most of the work was around making sure the --theater parameter checks the configured list of theaters properly. I didn't figure out how to get the acceptable list in the --help message, but I can live with that.

I also generalized calendar name and timezone parameters as per issue #6. Seattle and US/Pacific timezones remain the default. And now that I look at it, the prodid is already generated from the calendar name so I think that closes that one out.

I may ship this but I gotta get to work so I'll do a more comprehensive test run this weekend.