General cleanup
Mostly made the usage of self.base_url consistent; also tidied up setup.py a little bit.

Decided to leave the various per-theater variables as instance variables instead of class variables, mostly because self.theater is retrived from kwds and it just felt better to keep all the things I might need to change in one place. I flirted with the idea of pushing all the per-theater data into the config file but it'd take a bit of a refactor and doesn't really match how I think about the code, so I'm leaving it for now.

The other chunk of cleanup I want to do on this branch is modifying the duration of an event to be passed around as a timedelta.