Added Stella theater
The HTML is fairly bad for this theater (sorry!) -- there's a huge div tag with an attribute
containing JSON for all movie showings. Unfortunately that JSON uses double quotes, unescaped,
so the whole thing isn't really valid HTML. Thus this code is a bit messy because we have to
parse out the JSON by hand.

Once that was done it's fairly easy. I made a semi-arbitrary decision not to retrieve every
film page for the sake of figuring out the duration -- I'd rather reduce load on the theater's
Web page than be perfect.

Also added the theater to the config file.

Haven't tested this in a calendar yet; that's next.